More dead code cleanup - Removed most of the transaction-shim classes. Renamed file to extended-storage-transaction.ts, since that is the only remaining class.
